# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Aura Companion API
  version: 0.1.0
  description: API-first backend for a persistent AI companion (no physical integration).
servers:
  - url: http://127.0.0.1:8000
    description: Local development server
tags:
  - name: health
  - name: auth
  - name: conversation
  - name: memory
  - name: consent
  - name: observability
paths:
  /health:
    get:
      tags: [health]
      summary: Health check
      responses:
        '200':
          description: Service alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
  /v1/auth/token:
    post:
      tags: [auth]
      summary: Issue JWT bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/conversation/turn:
    post:
      tags: [conversation]
      summary: Process a conversation turn
      security:
        - BearerAuth: [conversation:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationRequest'
      responses:
        '200':
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/memory/store:
    post:
      tags: [memory]
      summary: Store a memory record
      security:
        - BearerAuth: [memory:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryRecord'
      responses:
        '200':
          description: Memory stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/memory/search:
    post:
      tags: [memory]
      summary: Search memories for a user
      security:
        - BearerAuth: [memory:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemorySearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemorySearchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/memory/purge:
    post:
      tags: [memory]
      summary: Purge user memory and issue a deletion receipt
      security:
        - BearerAuth: [memory:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurgeRequest'
      responses:
        '200':
          description: Purge receipt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurgeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/consent/{user_id}:
    get:
      tags: [consent]
      summary: Read user consent state
      security:
        - BearerAuth: [consent:write]
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Consent status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/consent/update:
    post:
      tags: [consent]
      summary: Update user consent state
      security:
        - BearerAuth: [consent:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentUpdateRequest'
      responses:
        '200':
          description: Updated status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/observability/events:
    get:
      tags: [observability]
      summary: List audit events
      security:
        - BearerAuth: [observability:read]
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        '200':
          description: Event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token with scoped claims.
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
  schemas:
    MessageIn:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
          minLength: 1
          maxLength: 5000
        modality:
          type: string
          enum: [text, voice, vision]
          default: text
        metadata:
          type: object
          additionalProperties:
            type: string
    ConversationRequest:
      type: object
      required: [user_id, session_id, message]
      properties:
        user_id:
          type: string
        session_id:
          type: string
        message:
          $ref: '#/components/schemas/MessageIn'
    SafetyResult:
      type: object
      required: [allowed, risk_level, reason, requires_break_reminder]
      properties:
        allowed:
          type: boolean
        risk_level:
          type: string
          enum: [low, medium, high]
        reason:
          type: string
        requires_break_reminder:
          type: boolean
    ConversationResponse:
      type: object
      required:
        - user_id
        - session_id
        - strategy
        - response_text
        - safety
        - memory_hits
        - persona_stable
        - generated_at
      properties:
        user_id:
          type: string
        session_id:
          type: string
        strategy:
          type: string
          enum:
            - cognitive_restructuring
            - behavioral_activation
            - mindfulness
            - emotional_support
            - self_monitoring
            - therapeutic_alliance
            - general
        response_text:
          type: string
        safety:
          $ref: '#/components/schemas/SafetyResult'
        memory_hits:
          type: array
          items:
            type: string
        persona_stable:
          type: boolean
        generated_at:
          type: string
          format: date-time
    MemoryRecord:
      type: object
      required: [user_id, session_id, role, content, retention_days, purpose, created_at]
      properties:
        user_id:
          type: string
        session_id:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        retention_days:
          type: integer
        purpose:
          type: string
          enum: [wellness, productivity, operations]
        created_at:
          type: string
          format: date-time
    MemorySearchRequest:
      type: object
      required: [user_id, query]
      properties:
        user_id:
          type: string
        query:
          type: string
    MemorySearchResponse:
      type: object
      required: [user_id, query, hits]
      properties:
        user_id:
          type: string
        query:
          type: string
        hits:
          type: array
          items:
            type: string
    ConsentUpdateRequest:
      type: object
      required: [user_id, gpc_enabled, one_click_opt_out, voice_enabled, vision_enabled]
      properties:
        user_id:
          type: string
        gpc_enabled:
          type: boolean
        one_click_opt_out:
          type: boolean
        voice_enabled:
          type: boolean
        vision_enabled:
          type: boolean
    ConsentStatusResponse:
      type: object
      required: [user_id, gpc_enabled, one_click_opt_out, voice_enabled, vision_enabled, updated_at]
      properties:
        user_id:
          type: string
        gpc_enabled:
          type: boolean
        one_click_opt_out:
          type: boolean
        voice_enabled:
          type: boolean
        vision_enabled:
          type: boolean
        updated_at:
          type: string
          format: date-time
    PurgeRequest:
      type: object
      required: [user_id, reason]
      properties:
        user_id:
          type: string
        reason:
          type: string
          enum: [user_request, retention_expired, compliance]
    PurgeResponse:
      type: object
      required: [user_id, deleted_records, receipt_id, deleted_at]
      properties:
        user_id:
          type: string
        deleted_records:
          type: integer
        receipt_id:
          type: string
        deleted_at:
          type: string
          format: date-time
    AuditEvent:
      type: object
      required: [event_type, details, created_at]
      properties:
        event_type:
          type: string
        user_id:
          type: string
          nullable: true
        details:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
    TokenRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in, scopes]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
        expires_in:
          type: integer
        scopes:
          type: array
          items:
            type: string